+++ packages/mantine/dev/main.tsx
@@ -1,4 +1,5 @@
 import { MantineProvider } from '@mantine/core';
+import '@mantine/core/styles.css';
 import * as React from 'react';
 import { createRoot } from 'react-dom/client';
 import { App } from 'react-querybuilder/dev';